refactor(sellproduct): 整理选择逻辑与名称访问
