+++ agent/go-service/sellproduct/goods/runtime_test.go
@@ -78,8 +78,8 @@ func TestOrderLocationPlanGroupsUsesStaticStrategy(t *testing.T) {
 // TestBuildRuntimeLocationPlanItemsSeparatesOutOfStock 验证只排除当前据点支持的缺货物品，并保持计划顺序。
 func TestBuildRuntimeLocationPlanItemsSeparatesOutOfStock(t *testing.T) {
 	groups := []itemPriorityGroup{
-		{ItemID: "item_a", DisplayName: "物品甲"},
-		{ItemID: "item_b", DisplayName: "物品乙"},
+		{ItemID: "item_a"},
+		{ItemID: "item_b"},
 	}
 	items, excludedOutOfStock, reserveSatisfied, excludedByUser := buildRuntimeLocationPlanItems(
 		groups,
@@ -87,11 +87,11 @@ func TestBuildRuntimeLocationPlanItemsSeparatesOutOfStock(t *testing.T) {
 		map[string]struct{}{"item_b": {}, "other_location_item": {}},
 		nil,
 	)
-	if len(items) != 1 || items[0].Name != "物品甲" || items[0].ReserveQuantity != 10 {
-		t.Fatalf("可售计划 = %+v，期望仅保留物品甲及其保留规则", items)
+	if len(items) != 1 || items[0].Name != "item_a" || items[0].ReserveQuantity != 10 {
+		t.Fatalf("可售计划 = %+v，期望仅保留 item_a 及其保留规则", items)
 	}
-	if len(excludedOutOfStock) != 1 || excludedOutOfStock[0] != "物品乙" {
-		t.Fatalf("缺货排除 = %v，期望仅包含当前据点的物品乙", excludedOutOfStock)
+	if len(excludedOutOfStock) != 1 || excludedOutOfStock[0] != "item_b" {
+		t.Fatalf("缺货排除 = %v，期望仅包含当前据点的 item_b", excludedOutOfStock)
 	}
 	if len(excludedByUser) != 0 {
 		t.Fatalf("用户排除 = %v，期望为空", excludedByUser)
@@ -103,8 +103,8 @@ func TestBuildRuntimeLocationPlanItemsSeparatesOutOfStock(t *testing.T) {
 
 func TestBuildRuntimeLocationPlanItemsSeparatesBlacklist(t *testing.T) {
 	groups := []itemPriorityGroup{
-		{ItemID: "item_a", DisplayName: "物品甲"},
-		{ItemID: "item_b", DisplayName: "物品乙"},
+		{ItemID: "item_a"},
+		{ItemID: "item_b"},
 	}
 	items, excludedOutOfStock, reserveSatisfied, excludedByUser := buildRuntimeLocationPlanItems(
 		groups,
@@ -112,14 +112,14 @@ func TestBuildRuntimeLocationPlanItemsSeparatesBlacklist(t *testing.T) {
 		map[string]struct{}{"item_a": {}},
 		nil,
 	)
-	if len(items) != 1 || items[0].Name != "物品乙" || items[0].ReserveQuantity != 20 {
-		t.Fatalf("可售计划 = %+v，期望仅包含物品乙", items)
+	if len(items) != 1 || items[0].Name != "item_b" || items[0].ReserveQuantity != 20 {
+		t.Fatalf("可售计划 = %+v，期望仅包含 item_b", items)
 	}
 	if len(excludedOutOfStock) != 0 {
 		t.Fatalf("黑名单物品不应被记为缺货：%v", excludedOutOfStock)
 	}
-	if !reflect.DeepEqual(excludedByUser, []string{"物品甲"}) {
-		t.Fatalf("用户排除 = %v，期望包含物品甲", excludedByUser)
+	if !reflect.DeepEqual(excludedByUser, []string{"item_a"}) {
+		t.Fatalf("用户排除 = %v，期望包含 item_a", excludedByUser)
 	}
 	if len(reserveSatisfied) != 0 {
 		t.Fatalf("用户黑名单物品不应记为已满足保留量：%v", reserveSatisfied)
@@ -128,8 +128,8 @@ func TestBuildRuntimeLocationPlanItemsSeparatesBlacklist(t *testing.T) {
 
 func TestBuildRuntimeLocationPlanItemsSeparatesSatisfiedReserve(t *testing.T) {
 	groups := []itemPriorityGroup{
-		{ItemID: "item_a", DisplayName: "物品甲"},
-		{ItemID: "item_b", DisplayName: "物品乙"},
+		{ItemID: "item_a"},
+		{ItemID: "item_b"},
 	}
 	items, excludedOutOfStock, reserveSatisfied, excludedByUser := buildRuntimeLocationPlanItems(
 		groups,
@@ -137,11 +137,11 @@ func TestBuildRuntimeLocationPlanItemsSeparatesSatisfiedReserve(t *testing.T) {
 		nil,
 		map[string]struct{}{"item_a": {}, "other_location_item": {}},
 	)
-	if len(items) != 1 || items[0].Name != "物品乙" {
-		t.Fatalf("可售计划 = %+v，期望仅包含物品乙", items)
+	if len(items) != 1 || items[0].Name != "item_b" {
+		t.Fatalf("可售计划 = %+v，期望仅包含 item_b", items)
 	}
-	if !reflect.DeepEqual(reserveSatisfied, []LocationPlanItem{{Name: "物品甲", ReserveQuantity: 10}}) {
-		t.Fatalf("保留量已满足 = %+v，期望包含物品甲及其保留数量", reserveSatisfied)
+	if !reflect.DeepEqual(reserveSatisfied, []LocationPlanItem{{Name: "item_a", ReserveQuantity: 10}}) {
+		t.Fatalf("保留量已满足 = %+v，期望包含 item_a 及其保留数量", reserveSatisfied)
 	}
 	if len(excludedOutOfStock) != 0 || len(excludedByUser) != 0 {
 		t.Fatalf("已满足保留量不应混入其他排除原因：缺货=%v 用户=%v", excludedOutOfStock, excludedByUser)
